bench(hc-core): criterion regression benches for field hot loops
Closes the gap the colleague flagged: "no Criterion-style regression
benches with stable inputs in CI — meaning the perf wins above can't
be tracked over time."

Four benches in `crates/hc-core/benches/field_arithmetic.rs`, each
keyed to a specific perf surface:

- `field_mul_scalar/single` — single-element mul. Tracks the
  Goldilocks fast-reduction primitive directly. A regression here
  proportionally slows every prover phase that touches field arithmetic.

- `mul_slices` — slice-wise mul, swept at n ∈ {256, 1024, 16384, 1M}
  to span the PAR_THRESHOLD=1024 transition. Catches both:
  (a) regressions in the per-element scalar path (the small sizes)
  (b) regressions in rayon dispatch / par_iter overhead (the threshold
  region)
  (c) regressions in the parallel-dominated regime (1M)

- `mul_assign_slices`, `add_assign_slices` — the in-place
  par_iter_mut paths, same size sweep. Distinct benches so a regression
  isolated to {add, mul} surfaces independently.

Inputs are seeded LCG-style so noisy hosts reproduce numbers across
runs; throughput labels (Melem/s) make `criterion --baseline`
comparisons meaningful for human review.

CI integration (3 new steps):
- compile-only step (`--no-run`) catches API drift in the bench file.
- `--quick --measurement-time 2` smoke run verifies the benches
  execute end-to-end on CI hardware. Total added CI time ~30s.
- Full baseline-comparison gating (regression-fail threshold) is
  deliberately deferred — that requires baseline storage and a
  comparison harness that's its own design problem.

Two pre-existing clippy lints fixed as a side effect:
- `redundant_closure` on `(1..=10).map(F::from_u64)` (was `|i| F::from_u64(i)`)
- `uninlined_format_args` on the `inverse of {v:?} failed` panic message
- `unnecessary_cast` on `EPSILON as u64` (already a u64)

Also gated `MODULUS_U128` to `#[cfg(test)]` since the fast-reduction
path no longer references it; only the test oracle does. Silences
`dead_code` under `clippy -D warnings`.

+++ crates/hc-core/src/field/prime_field.rs
@@ -8,6 +8,11 @@ use super::{FieldElement, TwoAdicField};
 
 /// Prime modulus `p = 2^64 - 2^32 + 1`.
 pub const GOLDILOCKS_MODULUS: u64 = 0xFFFFFFFF00000001;
+/// 128-bit form of the modulus. Only the test oracle uses it now (fast
+/// reduction lives entirely in u64 register-wide arithmetic). Kept
+/// available as a documented constant; gated to test builds to silence
+/// dead_code under `clippy -D warnings`.
+#[cfg(test)]
 const MODULUS_U128: u128 = GOLDILOCKS_MODULUS as u128;
 const MODULUS_MINUS_TWO: u64 = GOLDILOCKS_MODULUS - 2;
 /// `EPSILON = 2^32 - 1 = p mod 2^64`. Equivalent to `2^64 mod p`. Used as
@@ -74,7 +79,7 @@ impl GoldilocksField {
         let lo = t as u64; // bits 0..64 of t
         let hi_word = (t >> 64) as u64; // bits 64..128
         let hi = hi_word >> 32; // bits 96..128 of t, in [0, 2^32)
-        let mid = hi_word & EPSILON as u64; // bits 64..96 of t, in [0, 2^32)
+        let mid = hi_word & EPSILON; // bits 64..96 of t, in [0, 2^32)
 
         // Step 1: t1 ≡ lo - hi (mod p).
         // If `lo < hi` we underflow the u64; the wrapped value represents
